# docs/swagger.yaml
basePath: /api/users
definitions:
  model.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      passwordHash:
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервер с бд, для операций CRUD
  title: Repository
  version: "1.0"
paths:
  /api/users:
    get:
      consumes:
      - application/json
      description: Выводит список пользователей (можно указать ?limit=?offset=)
      parameters:
      - description: Сколько записей вернуть
        in: query
        name: limit
        type: integer
      - description: Смещение (сколько пропустить)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      summary: Список пользователей
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Добавляет нового пользователя
      parameters:
      - description: Параметры пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Ошибка валидации
          schema:
            type: string
      summary: Создание пользователя
      tags:
      - Users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: Ставит метку на удаление пользователя по переданному ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно удален
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      summary: Удаление пользователя по ID
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Выводит пользователя по переданному ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      summary: Получение пользователя по ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Обновляет данные о пользотеле на те, что передаются в запросе
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Параметры пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            type: string
      summary: Обновление данных о пользователе
      tags:
      - Users
swagger: "2.0"